backend/requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-decouple==3.8
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Валидация и работа с данными
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# Тестирование (опционально)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0